# frontend/moderator-dashboard-ui/docs/verification/backend/openapi_verified.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "B2B Platform API",
    "description": "API for B2B Platform - supplier moderation and parsing system",
    "version": "1.0.0"
  },
  "paths": {
...
    "/domains/queue": {
      "get": {
        "tags": [
          "Domains Queue"
        ],
        "summary": "List Domains Queue Endpoint",
        "description": "List domains queue entries with pagination.",
        "operationId": "list_domains_queue_endpoint_domains_queue_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "parsingRunId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parsingrunid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainsQueueResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/queue/{domain}": {
      "delete": {
        "tags": [
          "Domains Queue"
        ],
        "summary": "Remove From Domains Queue Endpoint",
        "description": "Remove domain from queue.",
        "operationId": "remove_from_domains_queue_endpoint_domains_queue__domain__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/pending": {
      "get": {
        "tags": [
          "Domains Queue"
        ],
        "summary": "List Pending Domains Endpoint",
        "description": "List unique pending domains (not in blacklist and not in suppliers).",
        "operationId": "list_pending_domains_endpoint_domains_pending_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingDomainsResponseDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/domains/pending/enrich": {
      "post": {
        "tags": [
          "Domains Queue"
        ],
        "summary": "Enrich Pending Domain",
        "description": "Run domain parser for a single pending domain and upsert supplier/Checko data.",
        "operationId": "enrich_pending_domain_domains_pending_enrich_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrichDomainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrichDomainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/domains/pending/clear": {
      "post": {
        "tags": [
          "Domains Queue"
        ],
        "summary": "Clear Pending Domains",
        "description": "Delete pending domains from domains_queue (they can reappear on next parsing runs).",
        "operationId": "clear_pending_domains_domains_pending_clear_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearPendingDomainsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClearPendingDomainsRequest": {
        "properties": {
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domains"
          }
        },
        "type": "object",
        "title": "ClearPendingDomainsRequest"
      },
...
      "DomainQueueEntryDTO": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "parsing_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsing Run Id"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Source of the URL: google, yandex, or both"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "keyword",
          "url",
          "status",
          "created_at"
        ],
        "title": "DomainQueueEntryDTO",
        "description": "Domain queue entry DTO."
      },
      "DomainsQueueResponseDTO": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DomainQueueEntryDTO"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "total",
          "limit",
          "offset"
        ],
        "title": "DomainsQueueResponseDTO",
        "description": "Response DTO for domains queue."
      },
...
      "EnrichDomainRequest": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          }
        },
        "type": "object",
        "required": [
          "domain"
        ],
        "title": "EnrichDomainRequest"
      },
      "EnrichDomainResponse": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inn"
          },
          "emails": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Emails",
            "default": []
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "status"
        ],
        "title": "EnrichDomainResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PendingDomainDTO": {
        "properties": {
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "occurrences": {
            "type": "integer",
            "title": "Occurrences"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "domain",
          "occurrences"
        ],
        "title": "PendingDomainDTO",
        "description": "Pending domain DTO (not in blacklist and not a supplier)."
      },
      "PendingDomainsResponseDTO": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/PendingDomainDTO"
            },
            "type": "array",
            "title": "Entries"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "total",
          "limit",
          "offset"
        ],
        "title": "PendingDomainsResponseDTO",
        "description": "Response DTO for pending domains list."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}